MELOSYS-2497 fikse queryparam i hentAktoerer. Samtidig stenge mulighenten til å slette aktøren Bruker

+++ service/src/main/java/no/nav/melosys/service/aktoer/AktoerService.java
@@ -8,6 +8,7 @@ import no.nav.melosys.domain.kodeverk.Aktoersroller;
 import no.nav.melosys.domain.kodeverk.Representerer;
 import no.nav.melosys.exception.FunksjonellException;
 import no.nav.melosys.exception.MelosysException;
+import no.nav.melosys.exception.TekniskException;
 import no.nav.melosys.repository.AktoerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
@@ -58,10 +59,13 @@ public class AktoerService {
         aktoerDto.setDatabaseID(aktoer.getId());
     }
 
-    public void slettAktoer(long databaseID) {
-        //det virker ikke å slette direkte på id-en.
-        Aktoer aktoer = new Aktoer();
-        aktoer.setId(databaseID);
+    public void slettAktoer(long databaseID) throws TekniskException, FunksjonellException {
+        Aktoer aktoer = aktørRepository.findById(databaseID).
+            orElseThrow(() -> new TekniskException("Klarte ikke slette aktøren. Fant ingen aktør på id: " + databaseID));
+
+        if (aktoer.getRolle().equals(Aktoersroller.BRUKER)) {
+            throw new FunksjonellException("Aktøren er en bruker. Det er ikke lov til å slette denne");
+        }
         aktørRepository.deleteById(aktoer);
     }
